[Cxx11]: Add 'resources'.
Resources, in this case, are dumped hex resources created with `xxd`.
The thinking here is to avoid having a heap of strings embedded in the
source code for things like `help' et al, but headers that contain
constant `unsigned char` arrays that can be pointed at.

This is a branch for the C++11 port of verbuild. Eventually it will replace the Qt version. Oh, when I say 'C++11', I mean 'C++14'.